chore: separate marketing groutp
