+++ src/routes/(marketing)/+layout.svelte
@@ -1,0 +1,7 @@
+<script lang="ts">
+  let { children } = $props();
+</script>
+
+<div class="mx-4 md:mx-16">
+  {@render children()}
+</div>
