+++ src/routes/(marketing)/+page.server.ts
@@ -1,4 +1,4 @@
-import type { PageServerLoad } from './$types';
+import type { PageServerLoad } from '../../../.svelte-kit/types/src/routes';
 
 export const load: PageServerLoad = async (event) => {
   return event.locals.info;
